fixed circular dependency and broke out gamesessions, merge works again now
